Admin & User routs with admin middleware

+++ routes/web.php
@@ -23,19 +23,17 @@ Route::get('/notification/error', 'NotificationController@error')->name('notific
 
 Auth::routes();
 
-Route::get('/users','UserController@index')->name('users.index');
-Route::get('/users/create','UserController@create')->name('users.create');
-Route::post('/users','UserController@store')->name('users.store');
-Route::get('/users/{user}','UserController@show')->name('users.show');
-Route::get('/users/{user}/edit','UserController@edit')->name('users.edit');
-Route::patch('/users/{user}','UserController@update')->name('users.update');
-Route::delete('/users{user}','UserController@destroy')->name('users.destroy');
+Route::get('/users','UserController@index')->name('users.index')->middleware(['auth','admin']);
+Route::get('/users/{user}','UserController@show')->name('users.show')->middleware(['auth','admin']);
+Route::get('/users/{user}/edit','UserController@edit')->name('users.edit')->middleware('auth');
+Route::patch('/users/{user}','UserController@update')->name('users.update')->middleware('auth');
+Route::delete('/users{user}','UserController@destroy')->name('users.destroy')->middleware(['auth','admin']);
 
 Route::get('/products','ProductController@index')->name('products.index');
-Route::get('/products/create','ProductController@create')->name('products.create');
-Route::post('/products','ProductController@store')->name('products.store');
+Route::get('/products/create','ProductController@create')->name('products.create')->middleware(['auth','admin']);
+Route::post('/products','ProductController@store')->name('products.store')->middleware(['auth','admin']);
 Route::get('/details/{product?}','ProductController@show')->name('products.show');
-Route::get('/products/{product}/edit','ProductController@edit')->name('products.edit');
-Route::patch('/products/{product}','ProductController@update')->name('products.update');
-Route::delete('/products{product}','ProductController@destroy')->name('products.destroy');
+Route::get('/products/{product}/edit','ProductController@edit')->name('products.edit')->middleware(['auth','admin']);
+Route::patch('/products/{product}','ProductController@update')->name('products.update')->middleware(['auth','admin']);
+Route::delete('/products{product}','ProductController@destroy')->name('products.destroy')->middleware(['auth','admin']);
